freeaskworld_connector/requirements.txt
websockets
numpy
Pillow
python-dotenv
aiortc